wrote test for utility function
